backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LangChain and AI
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.12
langchain-openai==0.0.2
openai==1.7.2

# Vector database
chromadb==0.4.22

# Document parsing
pypdf==3.17.4
docx2txt==0.8
python-docx==1.1.0

# Text processing
tiktoken==0.5.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0

# Monitoring and logging
prometheus-client==0.19.0
